refactor unit tests
* break into smaller files based on testing type
