+++ src/CMakeLists.txt
@@ -301,9 +301,18 @@ if(BUILD_TESTING)
     gen_func(square_and_disk_in_square)
   endif()
 
-  osh_add_exe(unit_tests)
-  set(TEST_EXES ${TEST_EXES} unit_tests)
-  test_func(run_unit_tests 1 ./unit_tests)
+  osh_add_exe(unit_math)
+  set(TEST_EXES ${TEST_EXES} unit_math)
+  test_func(run_unit_math 1 ./unit_math)
+  osh_add_exe(unit_array_algs)
+  set(TEST_EXES ${TEST_EXES} unit_array_algs)
+  test_func(run_unit_array_algs 1 ./unit_array_algs)
+  osh_add_exe(unit_mesh)
+  set(TEST_EXES ${TEST_EXES} unit_mesh)
+  test_func(run_unit_mesh 1 ./unit_mesh)
+  osh_add_exe(unit_io)
+  set(TEST_EXES ${TEST_EXES} unit_io)
+  test_func(run_unit_io 1 ./unit_io)
   osh_add_exe(r3d_test)
   test_func(run_test_r3d 1 ./r3d_test)
   osh_add_exe(mpi_tests)
